portofolioをts化
